add some method for admin
