+++ project/appweb/src/main/java/web/vasilizas/controller/person/TeacherSalary.java
@@ -1,4 +1,4 @@
-package web.vasilizas.controller;
+package web.vasilizas.controller.person;
 
 import javax.servlet.http.HttpServlet;
