+++ project/appweb/src/main/java/web/vasilizas/controller/person/TeacherSalaryController.java
@@ -1,5 +1,6 @@
-package web.vasilizas.controller;
+package web.vasilizas.controller.person;
 
+import vasilizas.myservice.person.TeacherService;
 
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
@@ -10,24 +11,25 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
 
-import static java.lang.Integer.parseInt;
-import static vasilizas.myservice.person.MyService.createAndAddPerson;
+import static java.lang.Double.valueOf;
 
-@WebServlet("/addteacher")
-public class TeacherController extends HttpServlet {
+@WebServlet("/teacher-salary")
+public class TeacherSalaryController extends HttpServlet {
+
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         String name = req.getParameter("name");
-        String age = req.getParameter("age");
+        String salary = req.getParameter("salary");
         String login = req.getParameter("login");
-        String password = req.getParameter("password");
 
+
+        TeacherService.setTeacherSalary(name, login, valueOf(salary));
         HttpSession session = req.getSession();
-        session.setAttribute("newteacher", name);
 
-        createAndAddPerson("Teacher", name, parseInt(age), login, password);
+        session.setAttribute("add", "You add new teacher " + name + " with salary " + salary );
 
-        RequestDispatcher requestDispatcher = req.getRequestDispatcher("/admin");
+
+        RequestDispatcher requestDispatcher = req.getRequestDispatcher("/admin/addpersonpar");
         requestDispatcher.forward(req, resp);
     }
 }
