+++ project/appweb/src/main/webapp/admin/addPerson.jsp
@@ -4,16 +4,22 @@
 <html lang=en>
 <head>
     <meta charset=UTF-8>
-    <title>Add Teacher parametr</title>
+    <title>Add person and parameter</title>
+    <style>
+        .text {
+            text-align:  center;
+        }
+    </style>
 </head>
 <body align="center">
+<jsp:include page="href.html"/>
 <h1 align="center">Work with teacher </h1>
 <h2 align="center"> Please enter age, name, login and password teacher </h2>
 <form action="/myweb/addteacher" method=post>
     <input required name=name placeholder=name type=text> <br/>
     <input required name=age placeholder=age type=text> <br/>
     <input required name=login placeholder=login type=text> <br/>
-    <input required name=password placeholder=password type=password> <br/>
+    <input required name=password placeholder=password type=text> <br/>
     <input type=submit value=Add!> <input type=reset value=Clean>
 </form>
 <p > </p>
@@ -21,17 +27,22 @@
 <form action="/myweb/teacher-salary" method=post>
     <input required name=name placeholder=name type=text> <br/>
     <input required name=login placeholder=login type=text> <br/>
-    <input required name=salary placeholder=salary type=password> <br/>
+    <input required name=salary placeholder=salary type=text> <br/>
     <input type=submit value=Add!> <input type=reset value=Clean>
 </form>
 <p > </p>
-<h2 align="center"> Please enter name, login and password student </h2>
+<h2 align="center"> Please enter name, age, login and password student </h2>
 <form action="/myweb/add-student" method=post>
     <input required name=name placeholder=name type=text> <br/>
     <input required name=age placeholder=age type=text> <br/>
     <input required name=login placeholder=login type=text> <br/>
-    <input required name=password placeholder=password type=password> <br/>
+    <input required name=password placeholder=password type=text> <br/>
     <input type=submit value=Add!> <input type=reset value=Clean>
 </form>
+<p > </p>
+<% if (session.getAttribute("add") == null ){
+       session.setAttribute("add", "There will result your work");} %>
+<h3> <%= session.getAttribute("add")%></h3>
+<% session.setAttribute("add",null); %>
 </body>
 </html>
