+++ project/appweb/src/main/webapp/admin/admin.jsp
@@ -7,7 +7,7 @@
 <body align=center>
 <jsp:include page="href.html"/>
 <h1 center> Admin work page </h1>
-<nav><a href="/myweb/admin/addteacherpar">Add teacher and salary</a> | <a href="/myweb/admin/avgsalary">Teacher's salary average</a>
+<nav><a href="/myweb/admin/addpersonpar">Add teacher and salary</a> | <a href="/myweb/admin/avgsalary">Teacher's salary average</a>
     | <a href="/myweb/admin/addstudent">Add student</a></nav>
 </body>
 </html>
