+++ project/appweb/src/main/webapp/admin/href.html
@@ -1,0 +1,2 @@
+<nav><a href="/myweb/work">Your work page</a> | <a href="/myweb/home">Go Home</a> | <a href="/myweb/about">Contact</a> |
+    <a href="/myweb/out">Sing out</a>  </nav>
