+++ project/appweb/src/main/webapp/student/student.jsp
@@ -9,12 +9,12 @@
 <h1 center> Your work page </h1>
 <h1 center> You can see your grades here. </h1>
 <h1 center>   </h1>
-<h1 center> Your name. </h1>
+<h2 center> Your name. </h2>
 <h1 center>   </h1>
-<h1 center> <%= session.getAttribute("name") %> </h1>
+<h2 center> <%= session.getAttribute("name") %> </h2>
 <h1 center>   </h1>
-<h1 center> Your grades. </h1>
+<h2 center> Your grades. </h2>
 <h1 center>   </h1>
-<h1 center> <%= session.getAttribute("marks") %> </h1>
+<h2 center> <%= session.getAttribute("marks") %> </h2>
 </body>
 </html>
